+++ src/main/java/lotto/util/validator/PurchaseLottoValidator.java
@@ -1,26 +1,28 @@
 package lotto.util.validator;
 
 import lotto.util.ExceptionMessage;
+import lotto.util.Util;
 
 public class PurchaseLottoValidator implements Validatable {
     @Override
-    public void validate(int purchase) throws IllegalArgumentException {
+    public void validate(String input) throws IllegalArgumentException {
+        int purchase = Util.convertStringToInt(input);
         validatePurchaseLottoRange(purchase);
         validatePurchaseLottoUnit(purchase);
     }
 
-    private static void validatePurchaseLottoRange(int purchase) {
+    private void validatePurchaseLottoRange(int purchase) {
         if (purchase < Range.MIN_PURCHASE_LOTTO.value) {
-            throw new IllegalArgumentException(ExceptionMessage.INVALID_LOTTO_PURCHASE_MIN_RANGE.getMessage());
+            throw new IllegalArgumentException(ExceptionMessage.INVALID_PURCHASE_MIN_RANGE.getMessage());
         }
         if (purchase > Range.MAX_PURCHASE_LOTTO.value) {
-            throw new IllegalArgumentException(ExceptionMessage.INVALID_LOTTO_PURCHASE_MAX_RANGE.getMessage());
+            throw new IllegalArgumentException(ExceptionMessage.INVALID_PURCHASE_MAX_RANGE.getMessage());
         }
     }
 
-    private static void validatePurchaseLottoUnit(int purchase) {
+    private void validatePurchaseLottoUnit(int purchase) {
         if (purchase % Range.LOTTO_UNIT.value != 0) {
-            throw new IllegalArgumentException(ExceptionMessage.INVALID_LOTTO_PURCHASE_UNIT.getMessage());
+            throw new IllegalArgumentException(ExceptionMessage.INVALID_PURCHASE_UNIT.getMessage());
         }
     }
