+++ src/main/java/lotto/util/validator/Validatable.java
@@ -1,5 +1,5 @@
 package lotto.util.validator;
 
 public interface Validatable {
-    void validate(int input) throws IllegalArgumentException;
+    void validate(String input) throws IllegalArgumentException;
 }
